bugfix version

+++ lib/lib_helper_apps.sh
@@ -1,36 +1,42 @@
 #!/bin/bash
 
-
-
 ### /etc/fail2ban/
 ###
 ###
 printf "\n\n***********************************************\n\nCopy helper apps to /etc/bitworker [y/n]: "
 if [ "$u_helper_apps" = "" ]; then
-        read u_helper_apps
+  read u_helper_apps
 fi
 
-
-    
 if [ "$u_helper_apps" = "y" ]; then
 
+  file004=/etc/bitworker
 
-        file004=/etc/bitworker
+  if [ ! -d "$file004" ]; then
+    mkdir $file004
+  fi
 
-        if [ ! -d "$file004" ]; then
-          mkdir $file004
-        fi
+  cp $u_path/files/bitworker/bw-* $file004/
+  chmod 700 $file004/bw-show-jails.sh
+  chmod 700 $file004/bw-unban-jails.sh
+  chmod 700 $file004/bw-import-mysql.sh
+  chmod 700 $file004/bw-list-zones.sh
+  chmod 700 $file004/bw-ftp.sh
 
-          cp $u_path/files/bitworker/bw-* $file004/
-          chmod 700 $file004/bw-show-jails.sh
-          chmod 700 $file004/bw-unban-jails.sh
-          chmod 700 $file004/bw-import-mysql.sh
-          chmod 700 $file004/bw-list-zones.sh
-          chmod 700 $file004/bw-ftp.sh
+  if [ ! -L "/bin/bw-show-jails.sh" ]; then
+    ln -s $file004/bw-show-jails.sh /bin/bw-show-jails.sh
+  fi
 
-          ln -s $file004/bw-show-jails.sh /bin/bw-show-jails.sh
-          ln -s $file004/bw-unban-jails.sh /bin/bw-unban-jails.sh
-          ln -s $file004/bw-import-mysql.sh /bin/bw-import-mysql.sh
-          ln -s $file004/bw-list-zones.sh /bin/bw-list-zones.sh
+  if [ ! -L "/bin/bw-unban-jails.sh" ]; then
+    ln -s $file004/bw-unban-jails.sh /bin/bw-unban-jails.sh
+  fi
+
+  if [ ! -L "/bin/bw-import-mysql.sh" ]; then
+    ln -s $file004/bw-import-mysql.sh /bin/bw-import-mysql.sh
+  fi
+
+  if [ ! -L "/bin/bw-list-zones.sh" ]; then
+    ln -s $file004/bw-list-zones.sh /bin/bw-list-zones.sh
+  fi
 
 fi
